+++ src/android/FileTransferBackground.java
@@ -1,19 +1,32 @@
 package com.spoon.backgroundfileupload;
 
+import static android.content.Context.POWER_SERVICE;
+
+import android.app.Activity;
+import android.app.Notification;
 import android.app.NotificationChannel;
 import android.app.NotificationManager;
+import android.content.BroadcastReceiver;
 import android.content.Context;
+import android.content.Intent;
+import android.content.IntentFilter;
 import android.content.pm.PackageManager;
+import android.net.ConnectivityManager;
+import android.net.Uri;
 import android.os.Build;
+import android.os.PowerManager;
+import android.provider.Settings;
 import android.util.Log;
 
 import androidx.annotation.Nullable;
+import androidx.annotation.RequiresApi;
 import androidx.work.BackoffPolicy;
 import androidx.work.Constraints;
 import androidx.work.Data;
 import androidx.work.ExistingWorkPolicy;
 import androidx.work.NetworkType;
 import androidx.work.OneTimeWorkRequest;
+import androidx.work.OutOfQuotaPolicy;
 import androidx.work.WorkInfo;
 import androidx.work.WorkManager;
 import androidx.work.WorkQuery;
@@ -56,6 +69,8 @@ public class FileTransferBackground extends CordovaPlugin {
 
     private static String currentTag;
     private static long currentTagFetchedAt;
+
+    public static NetworkReceiver networkReceiver = null;
 
     public void sendCallback(JSONObject obj) {
         /* we check the webview has been initialized */
@@ -323,6 +338,14 @@ public class FileTransferBackground extends CordovaPlugin {
     }
 
     private void startUpload(final String uploadId, final Data payload) {
+        // Create a BroadcastReceiver to check status of internet connectivity
+        if(networkReceiver == null) {
+            networkReceiver = new NetworkReceiver();
+            IntentFilter networkReceiverIntentFilter = new IntentFilter();
+            networkReceiverIntentFilter.addAction(ConnectivityManager.CONNECTIVITY_ACTION);
+            cordova.getContext().registerReceiver(networkReceiver, networkReceiverIntentFilter);
+        }
+
         Log.d(TAG, "startUpload: Starting work via work manager");
 
         OneTimeWorkRequest workRequest = new OneTimeWorkRequest.Builder(UploadTask.class)
@@ -334,6 +357,7 @@ public class FileTransferBackground extends CordovaPlugin {
                 .setBackoffCriteria(BackoffPolicy.LINEAR, 30, TimeUnit.SECONDS)
                 .addTag(FileTransferBackground.WORK_TAG_UPLOAD)
                 .addTag(getCurrentTag(cordova.getContext()))
+                .setExpedited(OutOfQuotaPolicy.RUN_AS_NON_EXPEDITED_WORK_REQUEST)
                 .setInputData(payload)
                 .build();
 
@@ -560,5 +584,20 @@ public class FileTransferBackground extends CordovaPlugin {
 
     public static void logMessage(String message) {
         Log.d("CordovaBackgroundUpload", message);
+    }
+
+    private class NetworkReceiver extends BroadcastReceiver {
+        @RequiresApi(api = Build.VERSION_CODES.O)
+        @Override
+        public void onReceive(Context context, Intent intent) {
+            ConnectivityManager connectivityManager = (ConnectivityManager) context.getSystemService(Context.CONNECTIVITY_SERVICE);
+            if((connectivityManager == null) || (connectivityManager.getActiveNetworkInfo() == null) || (connectivityManager.getActiveNetworkInfo().isConnectedOrConnecting() == false)) {
+                Log.d(TAG, "No internet connection");
+                UploadTask.UploadForegroundNotification.getRetryNotification(context);
+            } else {
+                NotificationManager notificationManager = (NotificationManager) context.getSystemService(Context.NOTIFICATION_SERVICE);
+                notificationManager.cancelAll();
+            }
+        }
     }
 }
